pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lmm"
version = "1.0.0"
description = "LMM - Classical QUBO Optimizer with Digital Dharma (no D-Wave required)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]
dharma = [
    "hnswlib>=0.8.0",
]
langchain = [
    "langchain-core>=0.2.0",
]
llamaindex = [
    "llama-index-core>=0.10.0",
]
server = [
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "httpx>=0.25",
]
all = [
    "hnswlib>=0.8.0",
    "langchain-core>=0.2.0",
    "llama-index-core>=0.10.0",
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "httpx>=0.25",
]

[project.urls]
Homepage = "https://github.com/nene3369/LMM"
Repository = "https://github.com/nene3369/LMM"

[project.scripts]
lmm = "lmm.cli:main"

[tool.setuptools.packages.find]
include = ["lmm*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
ignore = ["E402"]

[tool.ruff.lint.per-file-ignores]
"lmm/_vendor/**" = ["F401", "F811", "F841", "E741"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["lmm"]
omit = ["lmm/_vendor/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
show_missing = true
